Fixed the autologout problem
A logout was occurring when the menu page was clicked due to the fact that a period was used in the plugin name. Because the period is used for concatenation natively in PHP it is unsafe to use in that manner. I have hence changed all the "wp-code-highlight.js" to "wp-code-highlight-js".